Avoid goroutine leak and associated memory leak on specific case described below:
Is it possible than goroutine in charge of reading the req.replyCh channel will no longer waiting on this channel if:
 - sendTimeoutTimer & oneMoreSecondTimer timer is done;
 - incoming response message is arrived;
 - before removed requestID on the c.respMap.

So in this cas, goroutine launched by readResponseLoop will be forever blocked (it is goroutine leak) write into response.replyCh channel.

In order to avoid this situation, it is better to make buffer channel will accept always one write (even if the reader is no longer reading).